pyproject.toml
[project]
name = "mfa-webui"
version = "0.1.0"
description = "基于 Vue 的 通用 GUI。由 MaaFramework 强力驱动！"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.119.0",
    "maafw==4.5.6",
    "plyer>=2.1.0",
    "uvicorn==0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "imageio>=2.37.0",
    "nuitka>=2.8",
]
